fix: resolve circular dependency in getUserAvailability by extracting getPeriodStartDatesBetween utility
- Extract getPeriodStartDatesBetween into standalone utility function
- Remove dependency on getUserAvailabilityService from getBusyTimesFromLimits.ts
- Update UserAvailabilityService to delegate to utility function
- Breaks circular dependency chain while preserving all existing functionality
- Fixes ReferenceError during fast refresh in development mode
